# backend/src/main/java/com/socialmedia/auth/dto/LoginRequest.java
package com.socialmedia.auth.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class LoginRequest {

    @NotBlank
    private String usernameOrEmail;

    @NotBlank
    private String password;
}