# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Social Media API
  version: 1.0.0
  description: >
    REST API for the Social Media Clone application.
    All protected endpoints require a Bearer JWT token obtained via /auth/login or /auth/register.

servers:
  - url: http://localhost:8080
    description: Local development

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    # ── Auth ─────────────────────────────────────────────────
    RegisterRequest:
      type: object
      required: [username, email, password, displayName]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[a-zA-Z0-9_]+$'
          example: john_doe
        email:
          type: string
          format: email
          example: john@example.com
        password:
          type: string
          minLength: 8
          example: "Secret123!"
        displayName:
          type: string
          maxLength: 60
          example: John Doe

    LoginRequest:
      type: object
      required: [usernameOrEmail, password]
      properties:
        usernameOrEmail:
          type: string
          example: john_doe
        password:
          type: string
          example: "Secret123!"

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
        user:
          $ref: '#/components/schemas/UserResponse'

    # ── Users ────────────────────────────────────────────────
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        followersCount:
          type: integer
        followingCount:
          type: integer
        postsCount:
          type: integer
        createdAt:
          type: string
          format: date-time

    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
          maxLength: 60
        bio:
          type: string
          maxLength: 200
        avatarUrl:
          type: string

    # ── Posts ────────────────────────────────────────────────
    CreatePostRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 2000
          example: "Hello world! #firstpost"
        imageUrl:
          type: string
          example: "https://cdn.example.com/img/abc.jpg"
        privacy:
          type: string
          enum: [PUBLIC, FOLLOWERS_ONLY, PRIVATE]
          default: PUBLIC

    PostResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        content:
          type: string
        imageUrl:
          type: string
        privacy:
          type: string
          enum: [PUBLIC, FOLLOWERS_ONLY, PRIVATE]
        author:
          $ref: '#/components/schemas/UserResponse'
        likesCount:
          type: integer
        commentsCount:
          type: integer
        likedByCurrentUser:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    PagedPostsResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PostResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        last:
          type: boolean

    # ── Comments ─────────────────────────────────────────────
    CreateCommentRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000
          example: "Great post!"

    CommentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        content:
          type: string
        author:
          $ref: '#/components/schemas/UserResponse'
        postId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time

    PagedCommentsResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CommentResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        last:
          type: boolean

    # ── Follow ───────────────────────────────────────────────
    FollowStatusResponse:
      type: object
      properties:
        following:
          type: boolean
        followersCount:
          type: integer
        followingCount:
          type: integer

    # ── Errors ───────────────────────────────────────────────
    ApiError:
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time

# ── Paths ─────────────────────────────────────────────────────
paths:

  # Health
  /api/v1/health:
    get:
      tags: [Health]
      summary: Health check
      responses:
        '200':
          description: Service is up

  # ── Auth ──────────────────────────────────────────────────
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered; JWT returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: Username or email already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login and obtain JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Bad credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  # ── Users ─────────────────────────────────────────────────
  /api/v1/users/me:
    get:
      tags: [Users]
      summary: Get own profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user's profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated
    put:
      tags: [Users]
      summary: Update own profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /api/v1/users/search:
    get:
      tags: [Users]
      summary: Search users by username or display name
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Matching users
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserResponse'
                  page:
                    type: integer
                  totalElements:
                    type: integer
                    format: int64

  /api/v1/users/{username}:
    get:
      tags: [Users]
      summary: Get a user's public profile
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /api/v1/users/{username}/posts:
    get:
      tags: [Users]
      summary: Get posts by a specific user
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paged posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedPostsResponse'
        '404':
          description: User not found

  # ── Posts ─────────────────────────────────────────────────
  /api/v1/posts:
    post:
      tags: [Posts]
      summary: Create a post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '401':
          description: Not authenticated
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /api/v1/posts/feed:
    get:
      tags: [Posts]
      summary: Get home feed (posts from followed users + own posts)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated feed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedPostsResponse'
        '401':
          description: Not authenticated

  /api/v1/posts/{id}:
    get:
      tags: [Posts]
      summary: Get a post by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Post detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '404':
          description: Post not found
    delete:
      tags: [Posts]
      summary: Delete a post (owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Deleted
        '403':
          description: Not the post owner
        '404':
          description: Post not found

  /api/v1/posts/{id}/like:
    post:
      tags: [Posts]
      summary: Like a post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Post with updated like count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '409':
          description: Already liked
    delete:
      tags: [Posts]
      summary: Unlike a post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Post with updated like count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '409':
          description: Not liked yet

  # ── Comments ──────────────────────────────────────────────
  /api/v1/posts/{postId}/comments:
    post:
      tags: [Comments]
      summary: Add a comment to a post
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '404':
          description: Post not found
        '422':
          description: Validation error
    get:
      tags: [Comments]
      summary: Get comments for a post (oldest first)
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedCommentsResponse'

  /api/v1/posts/{postId}/comments/{commentId}:
    delete:
      tags: [Comments]
      summary: Delete a comment (comment author or post author only)
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Deleted
        '403':
          description: Not the comment or post owner
        '404':
          description: Comment or post not found

  # ── Follow ───────────────────────────────────────────────
  /api/v1/users/{username}/follow:
    post:
      tags: [Follow]
      summary: Follow a user
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Follow successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowStatusResponse'
        '403':
          description: Cannot follow yourself
        '409':
          description: Already following
    delete:
      tags: [Follow]
      summary: Unfollow a user
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Unfollow successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowStatusResponse'
        '409':
          description: Not following

  /api/v1/users/{username}/follow/status:
    get:
      tags: [Follow]
      summary: Check if current user follows a user
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: true if following, false otherwise
          content:
            application/json:
              schema:
                type: boolean

  /api/v1/users/{username}/followers:
    get:
      tags: [Follow]
      summary: List followers of a user
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated list of followers
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserResponse'

  /api/v1/users/{username}/following:
    get:
      tags: [Follow]
      summary: List users that a user is following
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated list of users being followed
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserResponse'